# src/TsaElectronics.Api/Models/AuthModels/AuthResponseModel.cs
namespace TsaElectronics.Api.Models.AuthModels;

public class AuthResponseModel
{
    public string Token { get; set; } = string.Empty;
    public Guid UserId { get; set; }
    public string Email { get; set; } = string.Empty;
    public string FirstName { get; set; } = string.Empty;
    public string LastName { get; set; } = string.Empty;
    public List<string> Roles { get; set; } = [];
}